[스프링 - 협업 미션] 포모(박은소리) 미션 제출합니다. (#39)
* feat: prod, test properties 추가

* refactor: prod를 mysql로 수정

* refactor: dataloader 수정

* refactor: prod, test properties 수정

* feat: 노선 전체 목록 조회 기능 추가
- 구간을 정렬하는 기능 추가

* refactor : 전체 노선 정보 조회 url lines/all -> lines/map 수정

* feat: 중복 아이디 체크 성공/실패 기능 구현

* refactor: 로그인 실패 시 에러 메세지를 반환하도록 수정

* refactor: 지하철 역 중복 시 예외처리 되도록 수정

* refactor: 지하철 역 수정 시 존재 여부 및 중복 이름에 대한 예외 처리 추가

* refactor: 지하철 역 삭제 시 존재 여부 체크

* refactor: 지하철 노선 생성/수정/삭제 시 예외 처리 기능 추가

* feat: swagger 기능 추가

* refactor: section 구간 추가/삭제에 대한 예외 구현

* refactor: path에 대한 예외 구현

* feat: 예상치 못한 에러에 대한 예외 처리

* fix: datatype converter 오류 수정

* test: HTTP status를 500 -> 400으로 수정

* refactor: 노선의 요금 필드 추가

* refactor: 노선의 요금 정책을 적용하도록 수정

* refactor: 노선별 요금에 대한 최대 추가 요금 정책 추가

* fix: 우선 배포를 위한 @valid 제거

* fix: 우선 배포를 위한 @valid 제거

* fix: 데이터베이스 재생성하도록 수정

* fix : 테스트 코드가 돌아가도록 수정

* refactor: 구간에 존재하는 지하철 역은 삭제할 수 없도록 추가

* refactor: 로그인/비로그인 사용자별, 나이별로 추가 요금 계산
- 비로그인 사용자는 성인 요금
- 로그인 사용자는 성인, 청소년, 어린이, 유아 요금이 계산됨

* refactor: 생성된 구간에 대한 정보를 응답에 담도록 수정

* refactor: 지하철 역 예외의 에러 메시지 동등성 비교

* refactor: 지하철 구간 입력 시 예외 메시지에 대한 세분화

* refactor: 프론트 요구 사항 반영 (/lines에 section 리스트 추가)

* refactor: exception 패키지 분리

* feat: Document 자동화

* refactor: section 제거 시 no content를 반환하도록 수정

* refactor: 코드 포맷팅 및 스웨거 삭제
- 필드 final 선언
- 클래스에게 상태를 물어보도록 하는 메서드 추가

* refactor: 수정된 라인을 반환하도록 변경

* feat: fare에 대한 패키지 및 Factory 생성

* refactor: dataloader의 색상을 프론트에서 사용가능하도록  변경

* refactor: Exception 핸들러 세분화

* refactor: 필드에 private final 추가

* refactor: transactional 추가

* refactor: 쿼리에 불필요한 new Object[] 제거

* refactor: distance에 대한 추상화

* refactor: 메소드 명을 더 명확히 하도록 수정

* refactor: 나이에 대한 Resolver를 Member Resolver에 통합

* refactor: 전체적인 exception 패키지 수정
- 도메인 구조 패키지 방식으로 exception 패키지 수정
- final, abstract로 하위 클래스와 상위 클래스 구분

* refactor: 클래스를 final로 변경

* refactor: fare 패키지 이동

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,29 +1,43 @@
 package wooteco.subway.station.application;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.line.dao.SectionDao;
+import wooteco.subway.line.exception.ExistStationInSectionException;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.exception.DuplicateStationException;
+import wooteco.subway.station.exception.NotFoundStationException;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 public class StationService {
-    private StationDao stationDao;
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
+        stationDao.findByName(stationRequest.getName())
+                .ifPresent(station -> {
+                    throw new DuplicateStationException();
+                });
+
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
     public Station findStationById(Long id) {
-        return stationDao.findById(id);
+        return stationDao.findById(id)
+                .orElseThrow(NotFoundStationException::new);
     }
 
     public List<StationResponse> findAllStationResponses() {
@@ -34,7 +48,30 @@ public class StationService {
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteStationById(Long id) {
+        findStationById(id);
+
+        if (sectionDao.countSectionByStation(id) > 0) {
+            throw new ExistStationInSectionException();
+        }
+
         stationDao.deleteById(id);
+    }
+
+    @Transactional
+    public StationResponse updateStationById(Long id, StationRequest stationRequest) {
+        findStationById(id);
+
+        stationDao.findByName(stationRequest.getName())
+                .filter(station -> station.sameNameAs(stationRequest.getName()) && !station.sameAs(id))
+                .ifPresent(station -> {
+                    throw new DuplicateStationException();
+                });
+
+        Station newStation = new Station(id, stationRequest.getName());
+        stationDao.update(id, newStation);
+
+        return StationResponse.of(findStationById(id));
     }
 }
